+++ packages/nimble-components/src/table/template.ts
@@ -40,6 +40,8 @@ export const template = html<Table>`
             ${when(x => x.columns.length > 0, html<Table>`
                 ${repeat(x => x.tableData, html<TableRowState>`
                     <${DesignSystem.tagFor(TableRow)}
+                        class="row"
+                        record-id="${x => x.id}"
                         :dataRecord="${x => x.record}"
                         :columns="${(_, c) => (c.parent as Table).columns}"
                     >